templated SoftLeptonTagInfo class
