+++ DataFormats/BTauReco/src/classes_def.xml
@@ -94,11 +94,9 @@
    <version ClassVersion="10" checksum="3825121604"/>
   </class>
   <class name="std::pair<edm::RefToBase<reco::Track>, reco::SoftLeptonProperties>"/>
-  <class name="reco::SoftLeptonTagInfo::LeptonMap"/>
+  <class name="reco::TemplatedSoftLeptonTagInfo<edm::RefToBase<reco::Track> >::LeptonMap"/>
 
-  <class name="reco::SoftLeptonTagInfo" ClassVersion="10">
-   <version ClassVersion="10" checksum="1042007752"/>
-  </class>
+  <class name="reco::SoftLeptonTagInfo" />
   <class name="reco::SoftLeptonTagInfoCollection"/>
   <class name="reco::SoftLeptonTagInfoRef"/>
   <class name="reco::SoftLeptonTagInfoFwdRef"/>
